Refactor sync
