--- choreographer/protocol/__init__.py
@@ -1,23 +1,0 @@
-"""choreographer.protocol provides classes and tools for Chrome Devtools Protocol."""
-
-from ._protocol import (
-    DevtoolsProtocolError,
-    Ecode,
-    ExperimentalFeatureWarning,
-    MessageTypeError,
-    MissingKeyError,
-    Protocol,
-)
-from ._session import Session
-from ._target import Target
-
-__all__ = [
-    "DevtoolsProtocolError",
-    "Ecode",
-    "ExperimentalFeatureWarning",
-    "MessageTypeError",
-    "MissingKeyError",
-    "Protocol",
-    "Session",
-    "Target",
-]
